requirements.txt
Django
djangorestframework
mysqlclient
django-mptt
jdatetime
django-jalali
